requirements.txt
Flask==3.1.0
google-genai==1.9.0
flask-cors